// frontend/src/Services/models/Event.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type Event = {
    id?: number;
    title?: string;
    conference?: string;
    conferenceAcronym?: string;
    eventAcronym?: string;
    start?: string;
    end?: string;
    categories?: Array<string>;
    tags?: Array<string>;
    description?: string;
    paperSubmission?: string;
    abstractSubmission?: string;
    notificationDue?: string;
    finalDue?: string;
    cameraReady?: string;
    deadlineNotes?: string;
    wikicfpUrl?: string;
    eventUrl?: string;
    submissionUrl?: string;
    location?: string;
    updatedAt?: string;
};